+++ CHANGELOG.md
@@ -5,9 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]: https://github.com/zombiezen/go-sqlite/compare/v1.0.0...main
+[Unreleased]: https://github.com/zombiezen/go-sqlite/compare/v1.1.0...main
 
-## [Unreleased][]
+## [1.1.0][] - 2024-01-14
+
+Version 1.1 introduces the ability to prepare connections on `sqlitex.Pool`,
+improves performance, and improves documentation.
+
+[1.1.0]: https://github.com/zombiezen/go-sqlite/releases/tag/v1.1.0
 
 ### Added